Functionalize downstream logic in classSuccessorMetadataEnvironment
Summary:
This diff follows up on the preceding diff by functionalizing (where by
functionalizing I mean decoupling from the dependency-tracked shared-memory
plubming and instead expressing functional dependencies directly via dependency
injection) all of the analysis logic in classSuccessorMetadataEnvironment.ml.

The previous diff functionalized the analysis logic that goes into the
produce_value function; this diff functionalizes all of the other functions
in the ReadOnly module.

As discussed in the preceding diff, the distinction between "upstream" and
"downstream" analysis logic is that upstream means whatever depends only on
upstream environments and is used to create the cache table, whereas
"downstream" means whatever depends on the cache table.

Both of the analysis logic modules live at the top, so we know that they are
decoupled from all of the Environment functor plumbing. In each case there
is a `Queries.t` record describing the dependencies used inside of the
analysis logic, and the environment modules are responsible for wiring up
dependencies correctly.

Reviewed By: grievejia

Differential Revision: D50892277

fbshipit-source-id: 0a9e2a3e6c41d64be3cabcb8e7a0a32c6e2f614d