Functionalize the upstream logic from ClassSuccessorMetadataEnvironment
Summary:
This diff is the beginning of an exploration of how to functionalize the
business logic that is currently intertwined with the dependency-tracked
shared-memory logic in our environment stack.

In this diff, I extract the *upstream* pure logic from
ClassSuccessorMetadataEnvironment into a separate module that is defined above (and is
therefore independent of) the environment itself.

The name UpstreamAnalysis is intended to capture an aspect of the refactor that I realized
is general: the business logic mixed into environment layers comes in two flavors that
need to be reflected in our refactor:
 - "upstream pure logic", which is analysis logic that depends on upstream layers and
   is used as part of the `produce_value` function. I call this "upstream" because it
   is upstream of the cache table.
 - "downstream pure logic", which is analysis logic baked into the ReadOnly modules that
   defines queries downstream layers can make use of. I call this "downstream" because
   it is downstream of the cache table.

Philosophical note that's relevant to this "upstream" vs "downstream" distinction:
- It's really just a matter of taste that we even have upstream pure logic - in
  principle these functions could be moved into the ReadOnly module of the
  *previous* layer. The way we have it laid out now is probably more
  human-friendly though since it makes sense for the logic to produce a cache
  value to live in the same file defining the cache table.
- The fact that it makes sense to colocate producing a value with the cache
  table is only true *because* we are intermingling the analysis logic and
  caching logic. When we abstract the caching from the analysis logic, we may
  be able to revisit the grouping of functionality.

Reviewed By: grievejia

Differential Revision: D50892000

fbshipit-source-id: 4d562119fd7b5096c352c76fc6fd2f293a7c4dfb