Make parsing interfaces equivalent
Summary:
Historically, we layered parsing logic in a way where the creation
of the cpython parser `~context` happened first, and then we passed
that context to a parsing function.

Note that the context wasn't being reused across modules, so this
layering just happened to be chosen, there's no efficiency reason
or other fundamental need for building the context at a higher
layer of the code.

But the errpy parser doesn't need context, and a result of this
layering combined with us not refactoring it later is that the errpy
and cpython parse functions wound up with inconsistent interfaces,
making the load_raw_source code more confusing to skim.

This diff pushes down the context building so that the interfaces
match for cpython and errpy parsing; that allows us to simply pick
the `parse` function we want to use.

Reviewed By: pradeep90

Differential Revision: D50977145

fbshipit-source-id: 29ed3deac06332f6baac0b4eaf8cc6327a2df694