Simplify parse success/failure types and logic
Summary:
Prior to this commit, the parsing code was more complex and customized
than really needed. We had a dedicated `parse_result` type that was
very similar to a `Result.t` except that it baked in the error variant,
and the `Error` arm of that type was identical to a `ParseResult.t`
except that it was missing the module path.

This was unnecessarily complex and confusing for a couple of resons:
- First, the `Error` variant of `parse_result` was shadowing the `Error`
  variant of `Result.t`, making things tricky to skim
- Second, we had to destrucure the `Error` and create a `ParserError`,
  but the module path was always available in the lower-level code anyway
- Third, the `create_parse_result` function was a little misnamed, since
  it specifically created an error as opposed to a `Success` but the name
  didn't indicate this.

To simplify, I've eliminated the `parse_result` type in favor of the isomorphic
`(Source.t, ParserError.t) result` (which is what the top-level code effectively
dealt with anyway) and pushed the construction of the `ParserError.t` down
instead of converting between an extra type.

Reviewed By: pradeep90

Differential Revision: D50979098

fbshipit-source-id: 855ce0ee6afbedd9e9570449fc05abdd1d5b4a6d