Extract a functional load_and_parse from load_raw_source
Summary:
This commit prepares AstEnvironment for extraction of an `UpstreamAnalysis`
module representing the parsing and phase-0 preprocessing in terms of
an abstracted `get_raw_source`, and separating that from the
`load_raw_source` outer-level function that is responsible for actually
storing the results in shared memory.

The next diff will pull all this functional analysis logic into a module
at the top, much as we did in D50892000 for the class successor metadata
environment.

Reviewed By: pradeep90

Differential Revision: D50981322

fbshipit-source-id: 1d8cd1739b44dbc4ca6c6330b7b7351049705f24