Extract repetitive boilerplate in parsing
Summary:
The CPython and Errpy parsers don't actually need to care
about module tracker possibly failing to load code - the differences
are just in how to actually parse a code string.

But for historical reasons we buried the call to ModuleTracker inside
of the parse functions, which led to a duplicat block for creating
errors if we failed to load code (e.g. because the file is unreadable).

This commit makes the parser-specific functions accept a code string as input
and pushes that block up to load_raw_source, making our logic more DRY.

Reviewed By: pradeep90

Differential Revision: D50977806

fbshipit-source-id: a30aa42e44028a7d1427286ae3142a1179f35100