Rename the pure-analysis modules
Summary:
In diff review of D50980076 Pradeep suggested alternative names
for `UpstreamAnalysis` and `DownstreamAnalysis`.

I tweaked his suggestions a little, but I agree that his proposed
names are a little better:
- `IncomingDataComputation` for `UpstreamAnalysis`
- `OutgoingDataComputation` for `DownstreamAnalysis`

The new names are probably more evocative of the role these functions plays
(they either operate on data "incoming" to the cache table or "outgoing" from
it), and more importantly they don't introduce puns with `UpstreamEnvironment`
or with the `Analysis` mega-package, instead using entirely unique names.

This diff applies the rename to both `AstEnvironment` and
`ClassSuccessorMetadataEnvironment`, which are the two layers we've
refactored in this way thus far.

Reviewed By: arthaud

Differential Revision: D51097905

fbshipit-source-id: 8a765a4d29949df500d2a996a25638a5ecc0f3be