Extract AstEnvironment.DownstreamAnalysis code
Summary:
The AstEnvironment downstream analysis logic is essentially the code
that deals with taking a raw source loaded from the parser and running
preprocess_phase1 on it, which includes wildcard expansion and can
thereby introduce nontrivial dependencies between UnannotatedGlobalEnvironment
and AstEnvironment.

This commit follows the example of D50892277 and functionalizes that code
in terms of an abstract `get_raw_source`-based interface, thereby decoupling
the analysis logic itself from the specifics of our environment stack and
from dependency tracking.

*Implementation note:*

This is one of the trickier modules to functionalize because unlike almost
all other layers, we actually bypass dependency tracking for get_raw_source
in some cases: because AstEnvironment -> UnannotatedGlobalEnvironment
updates are keyed 1:1 most of the time, we actually don't dependency track
the primary call from `get_processed_source` to `get_raw_source` that loads
the actual source we want to grab.

Instead, we only dependency track when we load *other* sources in order
to do wildcard expansion.

This saves us a bit of perf, because the dependency tracking logic is
not very efficient (we have to serialize triggers and send them from workers
to the main thread, which accounts for several percent of overall time
in typical incremental updates).

In order to hide this detail (which is entirely a concern of dependency
tracking, and irrelevant to the core business logic), we create a
`get_raw_source_and_maybe_track` inside of the `ReadOnly.get_processed_source`
that understands when it should bypass dependency tracking.

Luckily no other environment layer will need a hack like this :)

Reviewed By: aperez

Differential Revision: D50981319

fbshipit-source-id: e65bc5994d4d19437d092570bec83e570d46b795