Extract AstEnvironment.UpstreamAnalysis module
Summary:
As per D50892000, this diff extracts the functional business logic
of parsing code and handling errors into an `UpstreamAnalysis` module.
The module lives at the top of `AstEnvironment.ml`, and so we can be
sure that the interface does not depend on any of the shared memory
tables defined further down.

This refactor is a part of an initial move toward splitting the
storage and dependency tracking plumbing that powers our current
Pyre environment stack from the underlying analysis logic, which
will allow us to build alternate environment / cache table
implementations.

The immediate need for this is that we want to integrate Pyre with buck in a
way that will require new environment dumping and loading capabilities that
SharedMemory tables cannot support.

Reviewed By: pradeep90

Differential Revision: D50980076

fbshipit-source-id: c53736e711f1b3fe2821591a4feee9d15ac1dd6d