Batch type errors pull request for open files
Summary:
We would like to be able to batch type error requests for files, instead of needing to send separate requests for each file.

This diff creates two new **temporary** (and one permanent) request and response types for the codenav server that will allow us to migrate the API without breaking type errors for existing users.

- `NewGetTypeErrors` has the signature that we will eventually migrate `Request.GetTypeErrors` over to. In an upcoming diff, the signature of this will be copied to `GetTypeErrors`, and after that, this type will be removed.
- `NewTypeErrors` has the signature that we want to migrate `Response.TypeErrors.` to similarly to `NewGetTypeErrors`. It will also be removed.
- `TypeErrorsFileMapping` is a type that we'll be keeping, as it will become part of the signature for `TypeErrorsResponse`.

Reviewed By: stroxler

Differential Revision: D50901513

fbshipit-source-id: 02e3df1f2d2451d7db6daa75f4e78739e1986b87