Refactor AddTransform logic
Summary:
In order to reuse all of the logic for tracking name and finding
the parent class or module - which we will want to do for
"delete" and "replace" actions just as for "add" - this commit
extracts that boilerplate into the `PatchTransform` base
transform class.

This is a simple refactor based on inheritance and a smallish
code change; a better solution would actually be to use dependency
injection for transforming the parent body which is what I'll
do in the next diff.

Reviewed By: grievejia

Differential Revision: D51045161

fbshipit-source-id: 3c443317f5851d79e30cb0412cc49e0e066fbd2a