Add a transform for AddPatch
Summary:
# This stack: A typeshed patcher

This stack implements - at a level of correctness and completeness
that is adequate for all current patches, both against the pinned
upstream version and the latest - the typeshed patcher design as
outlined in our design doc from 2022:
https://fburl.com/gdoc/ju13p4v6
(sorry that link probably won't work externally)

This work is important for two reasons:
- It was already difficult to apply our patches, which were sensitive to line number and specific surrounding context, to upstream as of 2022; I remember it taking on the order of half a day to get the patches all regenerated then.
- Since then we've roughly 10x'ed the number of patches due to ReadOnly, meaning that by extrapolation it might have taken close to a week (for reference, it took me a week to implement this new patcher *and* apply the upgrade!)

**How much of the spec is implemented?**

It implements all the actions in the spec except for `DeleteImport`, which we realized we don't need in the current form:
- Many imports (for example `import foo` or `from foo import bar`) define a name in global scope, so a normal Delete or Replace action makes perfect sense. See D51161405 for details.
- The cases that don't fit cleanly are dotted imports (`import foo.bar`), in which case they technically don't define a single name but rather a stack of names because `foo` is implicitly imported and `bar` is an attribute on it. The spec doesn't actually help with this since the problem is more whether we're willing to use a qualified "name" or not as an action target, rather than the action per se.
   - For now we just don't support patching dotted names; it makes more sense to have that discussion if/when we need it.

**How general is the implementation?**

The implementation I wound up with is general enough to handle
all existing patches, both against our current pinned version of
typeshed and against the latest.

But it has multiple known gaps besides just the qualified-name
import I mentioned above. Most gaps are related to cases where
we need to be able to modify a top-level statement rather than
outright deleting or replacing it:
- imports defining many names like `from foo import bar, baz`
   can't currently be targeted for deletion or replacement
- if/else blocks are supported if they only define a single name,
   but if they define many names (for example a bunch of
  method definitions all in a single if block guarding on `sys.version`)
  they aren't currently supported.

The strategy for these unsupported cases is to detect them and raise
an error explaining what happened rather than silently fail to apply
a patch, in case an upgrade triggers one of these scenarios.

We could probably support this much better with a more
libcst-visitor-centric pattern, but we need to make sure it
is both simple enough to maintain and efficient enough for
dozens of patches on a huge file like `builtins.pyi` to be
fairly quick. I'll look into this later, but I want to prioritize an
upgrade first.

**About the stack itself**

This stack records my actual evolution of thinking about the implementation, so at some points the initial simpler design is generalized as I add more patching actions. I didn't try to rewrite
the stack so that the initial design becomes an oracle on future
needs.

I'm happy to take any suggestions, but anything that isn't trivial
I'd prefer to fix-forward rather than in-place, since the final
code is working and restacks can be quite difficult.

The code could probably benefit from unit tests of failure cases
as well; I didn't prioritize this because from my dev standpoint
I didn't need it (I was just recreating existing stubs) but I should
probably stack some on top.

# This diff: Start working on libcst transforms

I decided to start the patcher by working on the codemods themselves, which are easy to unit test, and only build the command-line tool once I have enough transforms ready to actually try patching at least one real file.

This is the first of the libcst transforms I will write; each transform
corresponds to one action from the typeshed patcher spec (which
describes how patch actions can be expressed in toml format).

The code should be mostly self-explanatory: we find the scope
the patch says to add to, then add statements to either the top
or bottom. The trickiest thing about it is that because LibCST
is indentation-aware class bodies have a bit of complexity in them
related to indented vs non-indented bodies; our result will always
be indented for simplicity.

**Next 2 diffs: refactor this code to make it reusable**

This diff is a fairly clean stand-alone implementation of an add
transform in idiomatic libcst, but it isn't actually the best code
layout given that we need a bunch of similar transforms each of
which operate on a parent scope.

As a result, I take this working code and refactor it in the next
2 diffs to really emphasize that the scope *finding* logic (which
is primarily a matter of traversing classes in the visitor) is
largely independent of the scope *patching* logic (which really
amounts to removing and/or adding some statements from/to
a class or module body).

Reviewed By: grievejia, connernilsen

Differential Revision: D50996596

fbshipit-source-id: 423807d1c9f8eef4e0a381db6e261ab424e2d553