Rename patch.py to patch_specs.py
Summary:
We've decided to be pretty strict about how we do imports in Pyre python code. We:
  (a) follow Google conventions of always importing modules
  (b) not using abbreviations, the way Google often does, for this scenario

There are some great advantages to this (it's super explicit and usually easy
to read), but as a result we ideally want to optimize our module names for
minimizing name collisions with variables in downstream code and avoiding
shadowing.

What do I mean by this? Naming a module like `patch` causes problems, because
we're going to want to name a variable `patch` in a ton of cases where we will
have to import this module. The name shadowing can then cause us to lose access
to the `patch` module in any function that has a variable named `patch`.

The easy fix is to make sure our module name is "module-like" rather than
"variable-like". Since I think we're going to have a lot of code that loops
over `patch` values, naming the module `patch_specs` will make it pretty
easy to avoid shadowing collisions.

I'm open to other names like `patch_types`, but I think calling it "specs" is
nice because "a patch" can mean many things depending on the context: the spec
data, the action of patching, the result of the patch, etc. Being explicit that
this module defines patch *specs*, i.e. the data declaring what patches we want
to run as opposed to actions or results of patching, seems clearer.

Reviewed By: grievejia

Differential Revision: D51045155

fbshipit-source-id: 7798c4dac019f70932664e07865ae4e16787fbb1