Add logic to pull upstream typeshed
Summary:
This diff begins wiring up plubming we need to actually build a
new typeshed patcher, which requires quite a bit more than just
implementing the patcher spec because at the very least we need:
- convenient ways to pull upstream typeshed
- decent UX for debugging a patch
- a cli for applying all the patches to a typeshed and writing to disk

In this diff I make a placeholder click command for fetching the
upstream typeshed, cleaning out irrelevant files, and dumping to a
directory.

Most of the logic is pulled from the old `download_typeshed.py`
script, but I simplified and cleaned up the logic a bit. We always
create a `Typeshed` first (which in a one-shot end-to-end script
we could just pass directly to patching), then write it out using
the `typeshed` helper function.

**Why am I not writing a zipfile?**

I decided to write to a directory rather than a zipfile because
we discovered last year that it is much better UX for people
hacking on patches to work with directories rather than zipfiles.

Why? We can just use editors, `cat`, and `diff` naively rather than
needing either dedicated tools or complicated chaining of unix
commands.

The extra overhead of always using directories is negligible so
I think it makes sense to always use directories when dumping to
disk.

Reviewed By: grievejia

Differential Revision: D51045158

fbshipit-source-id: cd00eb7fb61b586f4718881cdabb9b584b907da7