Rename argument to constructor
Summary:
Similar to my earlier name of `patch` to `patch_specs`, I think
there's an issue where we need to be a little careful of names
in order to keep our technical terms clear and crisp.

In the current code a "patch" always refers to a `path_specs.Patch`,
and I don't really want to then overload the same name for the
outcome of *applying* a patch.

As a result I want to rename the argument to `PatchedTypeshed` to
`patch_results`, to emphasize that this represents the outcome
of applying patch actions.

Reviewed By: grievejia

Differential Revision: D51045154

fbshipit-source-id: cef414f97d268840320ecab1fc1e2e32f086e107