Add a ReplaceTransform
Summary:
This is the last of the transforms that I think we need at the moment to
represent existing patches (DeleteImport is theoretically necessary but
as far as I know has no use cases yet).

As a result, I'd like to start working on getting the typeshed patcher
working end-to-end and figuring out what a good devx looks like on the
next diff; I can circle back and add DeleteImportTransform later since
it isn't necessary just yet.

Reviewed By: grievejia

Differential Revision: D51045162

fbshipit-source-id: dec9545fe03cea1edf42fc10cdc2f794e10e4c23