Tweak the typeshed.py interface
Summary:
I find the idea of a "file backed typeshed" and "writing to files" to
be a little less clear than "directory backed typeshed" and
"writing to a directory".

The reason is that a zipfile is also a file, so "file backed" doesn't
really seem as clear to me as "directory backed".

This diff just does a little rename.

Reviewed By: grievejia

Differential Revision: D51045156

fbshipit-source-id: 6ce76295e73ee3da37e39cd5cd9bdaaf22c93e99