Dependency-inject AddTransform logic
Summary:
This commit follows up on the previous refactor by using dependency
injection to make our AddTransform rely almost entirely on generic
code for
- tracking qualified names and finding the parent scope
- applying a plain-data transform to the parent scope

With this refactor, we only need one visitor which is helpful because
the visitor logic - which intermingles state tracking and data
transformations - is tricky to get right in my experience.

Plain data transforms, on the other hand, are relatively straightforward
(and also very familiar to our team from the OCaml world), so refactoring
into a state where the patch actions can be implemented as plain data
transforms that don't do anything except potentially add and/or remove
statements should make implementing the rest of the patch actions pretty easy.

Reviewed By: grievejia

Differential Revision: D51045159

fbshipit-source-id: 28ef19db983ebb1de98131e8a7dabf28a3d19e67