Wire up an endpoint to patch one file
Summary:
This commit adds a developer-facing command to the typeshed-patcher
cli that lets us patch a single stub from a typeshed we've previously
downloaded and extracted to a directory.

The results of the patch are printed as a diff, and we can optionally
write out the results (which is handy if the developer wants to
keep an editor window open to look at the results as they edit the
patches).

Using this new entrypoint, which is the first end-to-end flow from a
spec file to actual stubs, I also start building out a spec file
for our existing stubs.

Reviewed By: grievejia

Differential Revision: D51161400

fbshipit-source-id: 723baf45c6c7ced9c7ed6dde6d97c814e368f059