Dump content if we fail to parse
Summary:
Now that I can hack on the actual stub patches, it's time to fix UX issues I run
into as I go.

The first issue I had was a single character copy-paste error that led to invalid
syntax; the patcher crashes giving a syntax error with the line number but without
the content I have no easy way to figure out where in my patch spec I actually
messed up.

The easy fix is to just dump the content that failed to parse; for my use case
this was enough for me to very easily find the mistake and fix it.

Reviewed By: grievejia

Differential Revision: D51161403

fbshipit-source-id: a56a428cfe7dacb961bdcf99675b49a8c382ac69