pyre][typeshed_patcher] implement DeleteTransform
Summary:
This transform follows the same "modify the statement list
in a parent scope" logic that we are now using for AddTransform;
the difference is that we now need to simply find statements to
remove.

The key helper here is `matches_name`, which asks whether a
`libcst.BaseStatement` is associated with a particular
(simple, not qualified) name. We should be able to reuse
this helper in the replace action.

**(Incomplete logic)**

Note that we do not attempt to cover every case here - for
example, stubs can contain if blocks used to gate types on
the python version, and we don't handle this.

I'll need to fix this before actually rolling out the patcher, because
we do in fact replace an if/else block in the `enum.pyi` file. I'm
planning to fix-forward much later in the stack because I want
to get things working end-to-end before fine-tuning the
codemod logic.

Reviewed By: grievejia

Differential Revision: D51045152

fbshipit-source-id: a292b22e75e81be5640a6f95bb7c98bd69890c9c