Optimize: only use one visitor per stub
Summary:
The previous diff refactored PatchTransform so that actions are just
pure functions that mutate their parent scope (which is a sequence of
statements that we can add to / remove from / etc).

This diff uses that change by modifying PatchTransform so that we
can run patches in sequence in a single visitor.

The key insight is that after the refactor we did, the only real
change is to handle matching "whatever the current parent is" instead
of matching `self.parent`; this allows us to use a mapping of
patch actions and execute much, much faster

Reviewed By: grievejia

Differential Revision: D51161407

fbshipit-source-id: ae823482a2d7663190a40ae625861398a10dbe9b