Support matching names in simple if blocks
Summary:
It's somewhat common for stub files to define a name in an if block,
which allows things like python-version-dependent types or names that
only exist in certain python versions.

As a result, I've taken the conditional import logic support implemented
when skipping the import header in add actions, and refactored it to
be generic over any predicate.

This allows us to hook into the same logic in `statement_matches_name`,
so that we can handle if-blocks defining a single name in delete and
replace actions.

**Limitations**

Occasionally stubs will define many names in an if block, for example
a bunch of new methods defined in a recent version; this is pretty rare
and none of our stubs need it so for now so I haven't implemented it,
since that will require significantly more complex logic.

But if we see a case that looks like that come up - an if block where
some statements match the name and some don't - we will raise a
NotImplementedError rather than fail silently. This ensures that
we'll fix the sub patcher if and when we need to handle complex if
blocks, rather than accidentally regressing the stubs.

Reviewed By: grievejia

Differential Revision: D51161393

fbshipit-source-id: b60b5d01e8326ecfd4d2aff41ef650ed90f4e32a