Make `top` add actions work after import header
Summary:
When I was trying to recreate our existing `builtins.pyi` patched stub,
one of the problems I noticed is that my initial implementation of "top"
add actions added statements at the *very* top.

This is suitable for classes most of the time, but not usually suitable
for modules because we are typically trying to add to either the end of
the imports or above non-import statements.

For the moment I'm unaware of any cases where we need to add *above* all
imports, so I've modified `"top"` to simply skip over import statements,
and add above the first non-import statement.

There are some edge cases, for example import statements inside of an
`if` block (e.g. to gate on Python version); for now I'm not handling
that because none of our existing patches need it. It's fairly easy
to extend later.

Reviewed By: grievejia

Differential Revision: D51161394

fbshipit-source-id: 35829f3f2746f5a2a2ff43b9cc251e6d1d73d909