Refactor PatchTransform to prep for optimization
Summary:
I'm still not happy with the current devex of working on patches, it's too
slow to deal with a huge module like builtins.pyi that has lots of patches.

The only remaining obvious optimization is to stop visiting the entire
module many times - instead, we can group the patch actions (which are
always defined by mutations on a parent scope) into scopes and
update them all in one visitor.

This commit prepares for that by simpliflying the code so that we
can simply call a top-level `run_add_action` function that mutates
a parent scope (as a list of statements) for a particular action.

By setting that change up in this diff (which doesn't change any behaviors
at all, it just moves code around!) we can focus on the tricky changes
in the next diff.

Reviewed By: grievejia

Differential Revision: D51161397

fbshipit-source-id: 972a236f9e4e8a8bab3a1381a0b10c533dda6246