Support assignments in name matching
Summary:
This commit allows delete and replace actions to target assignment
statements (prior to this we only allowed defines, classes, and
annotated assignments).

I originally didn't think we needed this since stubs don't define
data, but type aliases happen via assignment so it is in fact
necessary; one of our patches to typeshed.pyi requires support for
this.

The logic is pretty simple; if we encounter any complex assignments
(the LHS of an assignment in Python can be a lot of things, e.g.
complicated multi-level patterns that might include indexes) we just
fail and dump an error message to be conservative; I don't really
expect stubs to make use of complex assignment targets.

Reviewed By: grievejia

Differential Revision: D51161398

fbshipit-source-id: 44bd048ef950c610355dd370126484eea03e41bd