Allow delete actions on imports of single names
Summary:
As of the previous diff, I have the existing typshed patches almost entirely
translated to the new patcher and spec.

The one exception is our patch to `os/__init__.py`, which requires
replacing these two lines
```
from . import path as _path
path = _path
```
with
```
from . import path
```

To do that I need a way to delete an import statement. The original
doc specified that we should use a DeleteImportAction for this,
but after talking with Jia I realized there's no point to that:
- if the import defines a single name (`import foo` or `from bar import foo`)
  then it makes sense to just use Delete for this
- if the import defines an attribute (`import foo.bar`), then the semantics
  of `name` wouldn't be clearly defined anyway; we can cross that bridge
  when we get there. One option would be to just allow dots in the `name`
  of an action and match the full attribute, but that's controversial
  because it's a bit like qualification.
  - Since we don't need dotted name deletion or replacement today, we
    can defer this discussion for now - we'll have to decide how to match
    dotted name imports if/when we wind up needing a patch for one.

With this change, I'm able to write the `os/__init__.py` patch as
a deletion plus a replacement.

**What isn't implemented?**

At this point there's no support for stripping just one import from
a list of imports in an import statement; if we run into that case
we will crash with an error. There's a similar issue that could
pop up if an if block defines many different names and we want to
delete just one of them.

Neither of these cases occurs in any existing patch, so for now
I'm deferring the work (with defensive code to catch and complain
loudly if this popws up in the future)

**If we need one of the unimplemented features, what do we do?**

The quick fix for this would be to allow users to mark a replace action
as "okay to delete extra things" and just recover the extra things
in the content.

That would be pretty easy - I think it could be implemented in a day
or so even by someone not too familiar with this code.

A prettier solution would be to actually be able to mutate the guts of
statements, but that is quite a lot tricker to implement and I would suggest
doing the hackier thing where we just delete extra names and put them back in
via a replace unless we see one of these edge cases come up repeatedly (the
fact that it has never happened yet makes me think it will likely be very
rare).

Reviewed By: grievejia

Differential Revision: D51161405

fbshipit-source-id: 3b174fb8d4942da85c079ae6af31d399355176c2