Improve the upstream filtering logic
Summary:
As noted in the last diff, the current result of pulling and
then patching upstream includes lots of extra config files and
directories that we don't actually want. The underlying problem
is I didn't correctly map 100% of the filtering code from
the old `download_typeshed.py` script into `upstream.py`.

In this commit I add the missing filters: ignore any file
that isn't in either `stubs` or `stdlib`.

Reviewed By: grievejia

Differential Revision: D51161408

fbshipit-source-id: 12b37effc40bc784495c73906066f95603329fe7