Patch asyncio.gather
Summary:
The latest upstream uses a generic form for fully-variadic
asyncio.gather calls (by fully-variadic I mean more than the
up-to-6 size calls that we support via overloads).

Unfortunately this is not going to work for us, for two reasons:
- Rolling it out might be a little tricky in places where there
  are bigger gather calls (I know for sure that hot instagram
  files have at least up to size 56!), because any callsites where
  the types of branches aren't actually joinable will produce
  type errors.
  - This isn't a reason why we *can't* roll out a change forcing
    gather calls to be well-typed, but we would need it to be a
    deliberate part of type coverage rather than an accidental
    side-effect of upgrade.
- More immediately, even trying to do this unification actually
  crashes Pyre - we enter a near-infinite loop (my guess is that
  it's polynomial complexity exploding rather than a "true"
  infinite loop) and eventually run out of memory when type checking
  distillery. We cannot roll this out without fixes to the Pyre
  expression resolution and constraint solving code, and it's not practical to
  block typeshed upgrade on such fixes.

This diff patches the code to pin it as the current form. I tried pinning my patch
to the new form which has size-6 calls, but that seems to also cause problems
so to unblock the upgrade I'm pinning it to the current code.

It's worth noting that we've discussed a type-level Map operator that would
express what `gather` actually does in a fully-variadic way; if we had that
and used it for `gather` then there'd be no problem, Pyre could analyze very
large calls tractably.

See D51138593 for the commands to regenerate typeshed from
the pinned version of upstream; I ran exactly the same commands
after tweaking the patch file to create this diff.

Reviewed By: pradeep90

Differential Revision: D51397092

fbshipit-source-id: 5c1185080e6c17827d0a681449dd9adee6c84678