Add toggle for diff view in patch-one-file
Summary:
Following up on a suggestion in D51161400. I suppose this will
make it finish a fraction of a second faster when diffs are off,
and it might be nice to have for files like `typing.pyi` where the diff
is so big that trying to read it isn't super helpful anyway.

I also eliminated my `---` boilerplate around the diff because the diff output
actually already has some markers that make it very easy to see where it starts and ends.

Reviewed By: kinto0

Differential Revision: D51450408

fbshipit-source-id: a48f5175fbf64548cd0bf2c9b84e0e01ef4d7166