Delete zip logic
Summary:
I've now finished the typeshed patcher and a full upgrade
cycle, and at no point was the zip logic useful.

Let's remove it. Zip files mattered when we actually stored
typesheds as zip files, but we got rid of that over a year ago because
it isn't good for devx (in particular, it's much harder to see diff
views, and debugging can involve piping zip operations to unix tools
which isn't very friendly).

Reviewed By: kinto0

Differential Revision: D51450406

fbshipit-source-id: c65ae97bbe7e9809f1d4759346c7fdaa48558f90