Update the patch-one-file interface
Summary:
This command is really just for developer experience, so that it's
quick to iterate on a single patch. It doesn't need to be fancy, since
the workflow will be write out the patch command and then use up arrow
to repeatedly run it and see the results.

This commit addresses feedback on D51161400 by simplifying the interface;
given the simple flow of the command we can just use stderr and stdout to
distinguish the debug information (exception traces and the diff) from
the output result.

It was suggested that I make the diff view lazy and toggleable.
I'll tackle that later in the stack (although also, it's a little hard
to imagine *not* wanting to be able to see the diff when iterating
on a patch file).

Reviewed By: kinto0

Differential Revision: D51450411

fbshipit-source-id: 5ce22c029143cf6f151e8a58543af1b1c5be3095