Use chained errors on parse failures
Summary:
The linter recommends this and it makes sense; I didn't fix it in my stack
because restacking is poorly-spent effort.

The actual suggestion was to also avoid printing `e.message`,
but I realized this is a bad idea: the order in which Python
dumps chained stack traces means that if you don't include
e.message in the top-level error message, it will be much harder to
actually find the line and column with the syntax error.

You can see in my example output where the location would show
up if I didn't dulicate `e.message` - often it wouldn't even be visible
in the terminal screen.

Still, it's nice to silence the lint error and surface the original
stack trace in case that might be helpful.

Reviewed By: kinto0

Differential Revision: D51450412

fbshipit-source-id: 9f5fe7ff6b51081e9324374809ee448072f57087